requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
anthropic==0.21.0
requests==2.31.0
pyyaml==6.0.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
gspread==5.12.0
python-multipart==0.0.6
httpx==0.25.2
pytz==2023.3
